+++ src/app/(dashboard)/admin/members/[memberId]/edit/page.tsx
@@ -1,0 +1,3 @@
+import AdminMemberEditPage from "@/src/components/pages/AdminMemberEditPage";
+
+export default AdminMemberEditPage;
