+++ src/app/(dashboard)/admin/organizations/[organizationId]/edit/page.tsx
@@ -1,0 +1,3 @@
+import AdminOrganizationEditPage from "@/src/components/pages/AdminOrganizationEditPage";
+
+export default AdminOrganizationEditPage;
