+++ src/app/(dashboard)/notices/page.tsx
@@ -1,3 +1,3 @@
-import noticesPage from "@/src/components/pages/NoticesPage";
+import NoticesPage from "@/src/components/pages/NoticesPage";
 
-export default noticesPage;
+export default NoticesPage;
