+++ src/app/(dashboard)/projects/[projectId]/edit/page.tsx
@@ -1,5 +1,3 @@
 import ProjectEditPage from "@/src/components/pages/ProjectEditPage";
 
-export default function Page() {
-  return <ProjectEditPage />;
-}
+export default ProjectEditPage;
