+++ src/app/(dashboard)/projects/[projectId]/questions/new/page.tsx
@@ -1,8 +1,6 @@
 // 질문글 작성 페이지
 
-import QuestionRegisterPage from "@/src/components/pages/QuestionRegisterPage";
+import ProjectQuestionsNewPage from "@/src/components/pages/ProjectQuestionsNewPage";
 import "./edit.css";
 
-export default function page() {
-  return <QuestionRegisterPage />;
-}
+export default ProjectQuestionsNewPage;
