+++ src/app/(dashboard)/projects/create/page.tsx
@@ -1,0 +1,5 @@
+// 프로젝트 생성 페이지
+
+import ProjectsCreatePage from "@/src/components/pages/ProjectsCreatePage";
+
+export default ProjectsCreatePage;
