--- src/app/(dashboard)/projects/new/page.tsx
@@ -1,7 +1,0 @@
-// 프로젝트 생성 페이지
-
-import ProjectRegisterPage from "@/src/components/pages/ProjectRegisterPage";
-
-export default function page() {
-  return <ProjectRegisterPage />;
-}
