+++ src/components/common/ArticleForm.tsx
@@ -10,7 +10,7 @@ import EditorJS from "@editorjs/editorjs";
 import ImageTool from "@editorjs/image";
 import FileAddSection from "@/src/components/common/FileAddSection";
 import LinkAddSection from "@/src/components/common/LinkAddSection";
-import SignUpload from "@/src/components/pages/ApprovalRegisterPage/components/SignUpload";
+import SignUpload from "@/src/components/pages/ProjectApprovalsNewPage/components/SignUpload";
 import DropDownInfoBottom from "./DropDownInfoBottom";
 
 const BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
