+++ src/components/pages/AdminMemberEditPage/index.tsx
@@ -1,0 +1,3 @@
+export default function AdminMemberEditPage() {
+  return <div>AdminMemberEditPage</div>;
+}
