+++ src/components/pages/AdminOrganizationEditPage/index.tsx
@@ -1,0 +1,3 @@
+export default function AdminOrganizationEditPage() {
+  return <div>AdminOrganizationEditPage</div>;
+}
