+++ src/components/pages/NoticesNewPage/index.tsx
@@ -8,7 +8,7 @@ import { NoticeRequestData } from "@/src/types";
 import { createNoticeApi } from "@/src/api/notices";
 import SelectInput from "@/src/components/common/FormSelectInput";
 import dynamic from "next/dynamic";
-import "./edit.css"
+import "./edit.css";
 
 const categoryData = [
   { id: 1, title: "서비스업데이트", value: "SERVICE_UPDATE" },
@@ -28,7 +28,7 @@ const ArticleForm = dynamic(
   },
 );
 
-export default function NoticeRegisterPage() {
+export default function NoticesNewPage() {
   const [title, setTitle] = useState<string>("");
   const [category, setCategory] = useState<string>("");
   const [priority, setPriority] = useState<string>("");
