+++ src/components/pages/ProjectsCreatePage/index.tsx
@@ -3,11 +3,10 @@
 "use client";
 
 import { Box, Flex } from "@chakra-ui/react";
-
 import BackButton from "@/src/components/common/BackButton";
-import ProjectForm from "@/src/components/pages/ProjectRegisterPage/components/ProjectForm";
+import ProjectForm from "@/src/components/pages/ProjectsCreatePage/components/ProjectForm";
 
-export default function ProjectRegisterPage() {
+export default function ProjectsCreatePage() {
   return (
     <Flex overflowX={"auto"}>
       <Box
